Allow teachers to choose to hide or allow grades.
When the instructor removes the "Grades" link from the course
navigation (in "Course Settings"), all grades are hidden from the
student. When included, the students are allowed to see/access
grades in that course. Note that the global "Grades" link is still
disallowed (since this would show grades for all courses).

Change-Id: I6aa7ff589001b5fa70c43286a705a2ce438b9b69
Reviewed-on: https://gerrit.instructure.com/5773
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Tom Metge <tom@instructure.com>