Add student activity report, grade export via API examples.
Student activity report can be used to aid in finding the student's
last day of attendence. The activities gathered can be customized
in the script itself, allowing institutions to only pull qualifying
activities.

Grade export is natively supported in Canvas but an example of
doing so via the API is included here just in case.

Change-Id: I0786b2b37fd243d75f9e4ccf9773975c2a35b4ca